+++ LightHreoProject/Assets/ReSources/Image.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aa483189b5a42ad40be6a9269a9ac3eb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
