+++ LightHreoProject/Assets/Scripts/GameManagerGen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fe01bf6040ed0cb49b80a0cd91cf5742
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
