+++ LightHreoProject/Assets/UpdateList(ReadMe).txt.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8d13aa708b8976c469829e3f4888adcd
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
